Register command completion via DI
